refactor with direct module imports
