+++ apps/antalmanac/src/backend/routers/notifications.ts
@@ -1,4 +1,10 @@
-import { RDS } from '$src/backend/lib/rds';
+import {
+    deleteAllNotifications,
+    deleteNotification,
+    retrieveNotifications,
+    updateAllNotifications,
+    upsertNotification,
+} from '$src/backend/lib/rds/notification';
 import { procedure, protectedProcedure, router } from '$src/backend/trpc';
 import { QuarterSchema, WebsocSectionStatusSchema, WebsocSectionTypeSchema } from '@packages/antalmanac-types';
 import { db } from '@packages/db';
@@ -21,14 +27,13 @@ const NotificationSchema = z.object({
     lastCodes: z.string(),
     notifyOn: NotifyOnSchema,
 });
-export type Notification = z.infer<typeof NotificationSchema>;
 
 const getStage = () => process.env.STAGE?.trim() || 'production';
 
 const notificationsRouter = router({
     get: protectedProcedure.query(async ({ ctx }) => {
         const stage = getStage();
-        return await RDS.retrieveNotifications(db, ctx.userId, stage);
+        return await retrieveNotifications(db, ctx.userId, stage);
     }),
 
     set: protectedProcedure
@@ -36,13 +41,13 @@ const notificationsRouter = router({
         .mutation(async ({ input, ctx }) => {
             const stage = getStage();
             await Promise.all(
-                input.notifications.map((notification) => RDS.upsertNotification(db, ctx.userId, notification, stage))
+                input.notifications.map((notification) => upsertNotification(db, ctx.userId, notification, stage))
             );
         }),
 
     updateNotifications: procedure.input(z.object({ notification: NotificationSchema })).mutation(async ({ input }) => {
         const stage = getStage();
-        await RDS.updateAllNotifications(db, input.notification, stage);
+        await updateAllNotifications(db, input.notification, stage);
     }),
 
     // Intentionally public: used by unauthenticated unsubscribe links
@@ -57,13 +62,13 @@ const notificationsRouter = router({
         )
         .mutation(async ({ input }) => {
             const stage = getStage();
-            await RDS.deleteNotification(db, input.userId, input.sectionCode, input.year, input.quarter, stage);
+            await deleteNotification(db, input.userId, input.sectionCode, input.year, input.quarter, stage);
         }),
 
     // Intentionally public: used by unauthenticated unsubscribe links
     deleteAllNotifications: procedure.input(z.object({ userId: z.string() })).mutation(async ({ input }) => {
         const stage = getStage();
-        await RDS.deleteAllNotifications(db, input.userId, stage);
+        await deleteAllNotifications(db, input.userId, stage);
     }),
 });
